Moved ProgressBar to UI
